Allow several spaces after LIMIT clause.
Previously the need to interpret bind variable as number was determined
based on presence of 'limit ?', 'Limit ?', 'lIMIT ?', 'LIMIT ?' in the
query. I.e. if there is indeed a LIMIT clause and it is followed by a
single whitespace and a bind variable placeholder. This makes it
impossible to write queries with multiple whitespaces after LIMIT clause
without explicit type binding.