Fix detection of INNODB engine in tests
When connecting to MariaDB server 10.1.x via MySQL client 5.5 then MySQL
client get information that server version is 5.5.5.

Variable have_innodb was removed in MySQL server 5.6 and is also not
available in MariaDB server 10.1. So INNODB engine is not detected.

So when possible use SHOW ENGINES which is supported since MySQL 4.1.2.